examples/postgresql-poll-trigger/requirements.txt
azure-functions
azure-functions-db[postgres]
